Move  all my Files into  _GithubPRs  Folder
to keep Original Proj Clean

----------

factor out my code - from axiel's view - as much as possible

min compatct std UMediList item , remove comments

animerepository remove comments.

remove comments whitespaces.

clean up comments UMListViev

app.kt minimize my code

Update CompactUserMediaListItem.kt

Update CompactUserMediaListItem.kt
clear compact imports

update gui    from UM_ListView with event.onItemSelected

it can now update gui    from UserMediaListView with event?.onItemSelected.

----------
it was really frustrating, until i found the exact event.

BUG-FIX: usermedilist item -  longTap wrong info bug fixed

if i call  AddNextAiringEpInfo_Compose inside UserMedialistView.kt

it opens WRONG dialog , SORT  button does wrong this. ALL ui event fucks up.

-------
safe bet is to move it back to anime repository.

1st app run it add 500-700ms extra time.
next time 1ms , cuz its cached in the KACHE.

fixed endless loop

found BUG -  coundt fix -  calling getairingnifo_compose  repeatedly.

i have 3 item in my list , now  there is a bug .

it calls compose again and again endless loop.

Update UserMediaListView.kt
--------------------
fix missing {,}  at AnilistQuery  -- caused deleting   getAiringxxxxxxSharedFlow func
