+++ app/src/main/java/com/axiel7/moelist/App.kt
@@ -8,7 +8,7 @@ import coil3.disk.DiskCache
 import coil3.disk.directory
 import coil3.memory.MemoryCache
 import coil3.request.crossfade
-import com.axiel7.moelist.Anilist.AnilistQuery
+import com.axiel7.moelist._GitHubPRs.Anilist.AnilistQuery
 import com.axiel7.moelist.data.model.media.TitleLanguage
 import com.axiel7.moelist.di.dataStoreModule
 import com.axiel7.moelist.di.databaseModule
@@ -46,7 +46,6 @@ class App : Application(), KoinComponent, KoinStartup, SingletonImageLoader.Fact
         )
 
         GlobalScope.launch(Dispatchers.IO) {
-            AnilistQuery.appThis = this;
             AnilistQuery.cache = AnilistQuery.New_ObjectKache()
         }
