+++ app/src/main/java/com/axiel7/moelist/_GitHubPRs/Anilist/Constants.kt
@@ -1,6 +1,6 @@
 @file:Suppress("unused")
 
-package com.axiel7.moelist.Anilist
+package com.axiel7.moelist._GitHubPRs.Anilist
 import com.axiel7.moelist.BuildConfig
 
 //package com.axiel7.anihyou.utils
@@ -12,7 +12,7 @@ const val ANILIST_URL = "https://anilist.co"
 const val ANILIST_API_URL = "$ANILIST_URL/api/v2"
 const val ANILIST_AUTH_URL = "$ANILIST_API_URL/oauth/authorize"
 const val ANIHYOU_AUTH_URL =
-    "${ANILIST_AUTH_URL}?client_id=${BuildConfig.ANILIST_CLIENT_ID}&response_type=token"
+    "$ANILIST_AUTH_URL?client_id=${BuildConfig.ANILIST_CLIENT_ID}&response_type=token"
 const val ANIHYOU_SCHEME = "anihyou"
 const val ANILIST_CALLBACK_URL = "$ANIHYOU_SCHEME://auth-response"
 const val ANILIST_GRAPHQL = "https://graphql.anilist.co/graphql"
