+++ app/src/main/java/com/axiel7/moelist/_GitHubPRs/Anilist/GlobalVariables.kt
@@ -1,4 +1,4 @@
-package com.axiel7.moelist.common
+package com.axiel7.moelist._GitHubPRs.Anilist
 
 class GlobalVariables(
     var accessToken: String? = null
