+++ app/src/main/java/com/axiel7/moelist/data/repository/AnimeRepository.kt
@@ -1,9 +1,7 @@
 package com.axiel7.moelist.data.repository
 
 import androidx.annotation.IntRange
-import com.axiel7.moelist.Anilist.AnilistQuery
-import com.axiel7.moelist.Anilist.AnilistQuery.Companion.AddNextAiringEpInfo_v2_withMeasureTime
-import com.axiel7.moelist.Anilist.secondsToDays
+import com.axiel7.moelist._GitHubPRs.Anilist.AnilistQuery
 import com.axiel7.moelist.data.model.Response
 import com.axiel7.moelist.data.model.anime.AnimeDetails
 import com.axiel7.moelist.data.model.anime.AnimeList
@@ -20,7 +18,6 @@ import com.axiel7.moelist.data.model.media.MediaStatus
 import com.axiel7.moelist.data.model.media.RankingType
 import com.axiel7.moelist.data.network.Api
 import io.ktor.http.HttpStatusCode
-import kotlin.system.measureTimeMillis
 
 class AnimeRepository(
     private val api: Api,
