+++ app/src/main/java/com/axiel7/moelist/ui/userlist/composables/MinimalUserMediaListItem.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.axiel7.moelist.Anilist.AiringEpN_in_Ndays_ToString
+import com.axiel7.moelist._GitHubPRs.Anilist.AiringEpN_in_Ndays_ToString
 import com.axiel7.moelist.R
 import com.axiel7.moelist.data.model.anime.AnimeNode
 import com.axiel7.moelist.data.model.anime.exampleUserAnimeList
@@ -80,7 +80,6 @@ fun MinimalUserMediaListItem(
                 if (isAiring) {
                     Text(
                         text = AiringEpN_in_Ndays_ToString(broadcast, item),
-                        //text = broadcast?.airingInString() ?: stringResource(R.string.airing),
                         color = MaterialTheme.colorScheme.primary,
                         fontSize = 16.sp,
                         lineHeight = 19.sp,
